feat: reset and change password endpoints
The password change expects the user to knwo their old password. Thus,
the old password has to be sent aswell as the new password and its
repeated entry for verification.

The password reset can be done without being logged into the account. An
email will be sent to the user's email with a link containing a token.
The endpoint for actually setting the new password needs the token and
the new password with its repeated entry.

The email part has yet to be implemented. For testing reasons, the token
is simply printed to the commandline.