feat!: change password change and reset system;
No need to send the repeated password to the backend. The check should
be done in the frontend.
For a password reset request, the email will now be needed as input in
the request body instead of the username.
For unarchive a simple auth login will be needed.
Instead of the custom login form, use the former used
OAuth2PasswordRequestForm. This messes the openapi-schema up, but a fix
will follow...