feat!: remodel reviews and review repr in loc info;
Reviews get a field `title` as string.
Renaming ReviewBase to ReviewApiIn and ReviewInfo to ReviewBase.

New model ReviewsSummary used in LocationDetailed, containing avg score,
review count and a list of (currently) up to 5 recent reviews.

Refactor in Location models:
- LocationBase gets fields name and trust_score
- LocationDetailed now inherits from LocationBase, not LocationShort
  - this needed some rewriting of the parsing between both classes
- Instead of a recent_reviews List, there is now a field `reviews` being
  a ReviewsSummary object

For each action on reviews (ReviewService) there is now a function in
the LocationService, that handles changes to the ReviewsSummary object.
Creating, updating and removing reviews from the `recent` list is thus
handled. The only missing case is, that the `recent` list is not filled
up when a review is deleted.