feat: verify user;
now, creating a new user marks the new user as archived for 20 minutes.
In that time, the user can be verified with an 8-digit code, which will
later be sent by email. For now, it is printed to the console.

If the user fails to verify in the first 20 minutes, the user object
will be deleted.

Currently, there is no process that cleans up unverified users
regularly.