feat!: reworking authentication w/ refresh tokens
Upon login, the user will now get an access token and a refresh token.
The access token will stay valid for the former 20 minutes. The refresh
token is used to generate a new access token and is itself valid for 30
days. Only the refresh token is saved in the database thus reducing
frequent checks for the validity of the access token.